pyproject.toml
[tool.poetry]
name = "test_eapteka"
version = "0.1.0"
description = ""
authors = ["Filipp Chachin"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^6.2.4"
allure-pytest = "^2.9.43"
loguru = "^0.5.3"
selenium = "^3.141.0"
pytest-rerunfailures = "^10.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
